working the way I want it to, divvy up the files
